Add testfixture function to make testing efficient
function `testfixture` receives the function address and the address of
the input/golden data, then compare the function output and the golden
data.

Since some functions receive 1 argument and the other functions receive
2 arguments, `testfixture` also needs an arguments that determines the
number of arguments of the testing function.

It is useful to test a variety of arithmetic functions in bfloat16.